# Conflicts:
#	CUE4Parse/FileProvider/Vfs/AbstractVfsFileProvider.cs

+++ CUE4Parse/UE4/Assets/Exports/Wwise/FWwiseEventCookedData.cs
@@ -8,18 +8,18 @@ namespace CUE4Parse.UE4.Assets.Exports.Wwise;
 [StructFallback]
 public readonly struct FWwiseEventCookedData
 {
-    public readonly int EventId;
-	public readonly FWwiseSoundBankCookedData[] SoundBanks;
-	public readonly FWwiseMediaCookedData[] Media;
-	public readonly FWwiseExternalSourceCookedData[] ExternalSources;
-	public readonly FWwiseSwitchContainerLeafCookedData[] SwitchContainerLeaves;
-	public readonly UScriptSet RequiredGroupValueSet; // FWwiseGroupValueCookedData[]
-	public readonly EWwiseEventDestroyOptions DestroyOptions;
-	public readonly FName DebugName;
+    public readonly uint EventId;
+    public readonly FWwiseSoundBankCookedData[] SoundBanks;
+    public readonly FWwiseMediaCookedData[] Media;
+    public readonly FWwiseExternalSourceCookedData[] ExternalSources;
+    public readonly FWwiseSwitchContainerLeafCookedData[] SwitchContainerLeaves;
+    public readonly UScriptSet RequiredGroupValueSet; // FWwiseGroupValueCookedData[]
+    public readonly EWwiseEventDestroyOptions DestroyOptions;
+    public readonly FName DebugName;
 
     public FWwiseEventCookedData(FStructFallback fallback)
     {
-        EventId = fallback.GetOrDefault<int>(nameof(EventId));
+        EventId = (uint)fallback.GetOrDefault<int>(nameof(EventId));
         SoundBanks = fallback.GetOrDefault<FWwiseSoundBankCookedData[]>(nameof(SoundBanks), []);
         Media = fallback.GetOrDefault<FWwiseMediaCookedData[]>(nameof(Media), []);
         ExternalSources = fallback.GetOrDefault<FWwiseExternalSourceCookedData[]>(nameof(ExternalSources), []);
